+++ BovineLabs.Core.Extensions/LifeCycle/DestroyEntity.cs
@@ -1,15 +1,15 @@
-﻿// <copyright file="EntityDestroy.cs" company="BovineLabs">
+﻿// <copyright file="DestroyEntity.cs" company="BovineLabs">
 //     Copyright (c) BovineLabs. All rights reserved.
 // </copyright>
 
-#if !BL_DISABLE_DESTROY
-namespace BovineLabs.Core.Destroy
+#if !BL_DISABLE_LIFECYCLE
+namespace BovineLabs.Core.LifeCycle
 {
     using Unity.Entities;
 
     /// <summary> Unified destroy component allowing entities to all pass through a singular cleanup group. </summary>
     [ChangeFilterTracking]
-    public struct EntityDestroy : IComponentData, IEnableableComponent
+    public struct DestroyEntity : IComponentData, IEnableableComponent
     {
     }
 }