+++ BovineLabs.Core.Extensions/LifeCycle/DestroySystemGroup.cs
@@ -2,13 +2,13 @@
 //     Copyright (c) BovineLabs. All rights reserved.
 // </copyright>
 
-#if !BL_DISABLE_DESTROY
-namespace BovineLabs.Core.Destroy
+#if !BL_DISABLE_LIFECYCLE
+namespace BovineLabs.Core.LifeCycle
 {
     using Unity.Entities;
 
-    [UpdateAfter(typeof(EndSimulationEntityCommandBufferSystem))]
-    [UpdateInGroup(typeof(SimulationSystemGroup), OrderLast = true)]
+    [UpdateAfter(typeof(BeginSimulationEntityCommandBufferSystem))]
+    [UpdateInGroup(typeof(SimulationSystemGroup), OrderFirst = true)]
     public partial class DestroySystemGroup : ComponentSystemGroup
     {
     }