+++ BovineLabs.Core/Collections/CollectionCreator.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="ContainerCreator.cs" company="BovineLabs">
+﻿// <copyright file="CollectionCreator.cs" company="BovineLabs">
 //     Copyright (c) BovineLabs. All rights reserved.
 // </copyright>
 
@@ -24,6 +24,24 @@ namespace BovineLabs.Core.Collections
             where TValue : unmanaged
         {
             var allocator = hashMap->m_Data.Allocator;
+            hashMap->Dispose();
+            AllocatorManager.Free(allocator, hashMap);
+        }
+
+        public static UnsafeMultiHashMap<TKey, TValue>* CreateMultiHashMap<TKey, TValue>(int initialCapacity, AllocatorManager.AllocatorHandle allocator)
+            where TKey : unmanaged, IEquatable<TKey>
+            where TValue : unmanaged
+        {
+            var hashMap = AllocatorManager.Allocate<UnsafeMultiHashMap<TKey, TValue>>(allocator);
+            *hashMap = new UnsafeMultiHashMap<TKey, TValue>(initialCapacity, allocator);
+            return hashMap;
+        }
+
+        public static void Destroy<TKey, TValue>(UnsafeMultiHashMap<TKey, TValue>* hashMap)
+            where TKey : unmanaged, IEquatable<TKey>
+            where TValue : unmanaged
+        {
+            var allocator = hashMap->data.Allocator;
             hashMap->Dispose();
             AllocatorManager.Free(allocator, hashMap);
         }